requirements.txt
requests>=2.27.1
numpy>=1.14.0
opencv-python>=4.5.5.64
Flask==3.0.3
Flask-Cors==4.0.1
pillow==10.3.0
face-recognition==1.3.0
face_recognition_models
Click>=6.0
dlib>=19.3.0
scipy>=0.17.0
fdk>=0.1.21